+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Addition of payment\_control and payments attributes
     * Addition of automatic\_payments? and manual\_payments? methods
     * Addition of is\_attending?, is\_not\_attending?, is\_invited? and is\_shown? methods
+    * update(api, field\_id) method now updates only changed fields, becoming update(api)
   * Add :network Section type
 
 ## Version 0.3.0